ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.0
pandas==2.2.0
scikit-learn==1.5.2
joblib==1.3.2
python-dotenv==1.0.0
pymongo==4.5.0
requests==2.31.0
